ADD: test_model
At this stage, the network learns how to play with random start position
in a static environment. But as soon as the positions of goal, trap and wall
change, the network fails. This is caused by "catastrophic forgetting". In
simple words, it means learning new stuff overwrites the stuff learned before.
Therefore, it causes instability in learning.

To test it, run the python script once to train the network. After
dql_model.pt has been created in the local path, run the script again
to see the test result. Running the script multiple times to see how
well it performs with random player position.